+++ src/main/java/net/fleshz/block/WoodRack.java
@@ -24,7 +24,6 @@ import net.minecraft.state.property.Properties;
 import net.minecraft.util.ActionResult;
 import net.minecraft.util.BlockMirror;
 import net.minecraft.util.BlockRotation;
-import net.minecraft.util.Hand;
 import net.minecraft.util.ItemScatterer;
 import net.minecraft.util.hit.BlockHitResult;
 import net.minecraft.util.math.BlockPos;
@@ -37,7 +36,6 @@ import net.minecraft.world.WorldView;
 import net.minecraft.block.ShapeContext;
 import org.jetbrains.annotations.Nullable;
 
-@SuppressWarnings("deprecation")
 public class WoodRack extends Block implements BlockEntityProvider {
 
     public static final VoxelShape SHAPENORTH;
@@ -64,7 +62,7 @@ public class WoodRack extends Block implements BlockEntityProvider {
     }
 
     @Override
-    public ActionResult onUse(BlockState state, World world, BlockPos pos, PlayerEntity player, Hand hand, BlockHitResult hit) {
+    protected ActionResult onUse(BlockState state, World world, BlockPos pos, PlayerEntity player, BlockHitResult hit) {
         WoodRackEntity woodReckEntity = (WoodRackEntity) world.getBlockEntity(pos);
         ItemStack stack = woodReckEntity.getStack(0);
         if (stack.isEmpty()) {