+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ### Added:
 - Added adventurez, meadow, naturalist compat
 ### Fixed:
-- 
+- Fixed hopper interaction
 ### Changed:
 -