+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### Added:
-- Added adventurez, meadow, naturalist compat
+- 
 ### Fixed:
 - Fixed trader llama drop
+- Fixed naturalist compat
 ### Changed:
 -